Move ray to actorlearner
