+++ adept/container/__init__.py
@@ -14,6 +14,6 @@
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 from .local import Local
 from .distrib import DistribHost, DistribWorker
-from .actorlearner import ActorLearnerHost, ActorLearnerWorker
+from .actorlearner import ActorLearnerHost, ActorLearnerPeer, ActorLearnerWorker
 from .evaluation import EvalContainer
 from .init import Init
