+++ adept/container/actorlearner/__init__.py
@@ -1,0 +1,2 @@
+from .learner_container import ActorLearnerHost, ActorLearnerPeer
+from .rollout_worker import ActorLearnerWorker
